verneuil/linuxvfs: implement xFullPathname
Summary:
FullPathname is probably the most annoying VFS method to implement on
unix, because sqlite's model doesn't quite fit with the unix way:
sqlite must be able to create journal files next to db files, and
know that every process that accesses the db file will look for the
journal in the same place.  That doesn't even make sense with
hardlinks, but sqlite does try to do something useful with symlinks.

When a file exists, we can simply rely on realpath(3).  However,
that's not useful when, e.g, a symlink refers to a file that doesn't
exist yet.

When realpath(3) fails, we must instead:

1. readlink(2) to figure out where a symlink point
2. repeat #1 until the path isn't a symlink
3. resolve the parent directory with realpath(3), append the filename.

We can't pass the fully un-symlinked path to `realpath` in step 3
because we already know that some of the path's components don't
exist.  In the common case, only the target file is missing (will be
created by the open(2) call), and resolving the parent directory
gives us something useful.

When part of the parent directory path is also missing, any open(2)
call will fail, so it doesn't seem too important to figure out what to
do.  For now, the VFS uses the parent path as is when realpath(3)
fails with ENOENT or EACCES: EACCES might not actually be a problem
for open, and ENOENT will just fail later if we try to open that path.

What's important is that, when the path refers to a symlink, we
fully chase down the symlink references (or error out): this
guarantees that the journal file will live next to the actual
symlink target file, and not the symlink itself.

Part of Txxxx.

Test Plan: shimvfs/test.sh

Reviewers: wandrews, ahicks, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx