verneuil/linuxvfs: implement HAS_MOVED fcntl
Summary:
We now track the backing path for shim files, and store the file's
device and inode id when we opened it (it would be ideal to use fstat,
but we can't do that while we still rely on sqlite's default VFS
for file operations).

Regardless, this lets us detect when named files have moved, by
stat-ing the same path when asked, and comparing the new device and
inode.

Future work: close the race by fstat-ing the open file, and either
remember the name of temporary files, or make sure they never have
a real filename at all.

Part of Txxxx.

Test Plan: shimvfs/test.sh

Reviewers: wandrews, sbahra, ahicks

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx