verneuil/linuxvfs: fake test-only fsync counts more accurately
Summary:
We want to keep track of the first time we fdatasync a journal file:
when that happens, sqlite tests expect us to guarantee the file is
visible after a crash by fsyncing the parent directory.  We don't
fsync directories, but we do increment test-only counters to show
that we correctly approximate the VFS interface.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: This is test-only. more sqlite tests pass.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: sbahra

Differential Revision: https://foo.backtrace.io/Dxxxx