NFC verneuil/vfs: copy the Linux VFS to `vfs.c`
We'll want to rename and modify some functions, but the bulk of the
code is identical.  Start with a 1:1 copy to make the changes easier
to review.

We want to keep the Linux VFS because it will be useful for testing if
we ever implement more functionality in the VFS (e.g., file
preallocation).

Part of Txxxx.