verneuil/vfs_ops: intercept lock and unlock calls
Summary:
And confirm that the Rust VFS's write methods are only invoked on
write-locked files.

This gives us a hook to snapshot the contents of a DB file when we
know it is valid: whenever sqlite wants to releases its lock on a db
file.  We also downgrade any exclusive lock to a plain shared (reader)
lock before the hook code, so the snapshot will block writes
(necessary), but not readers or write transactions that have
yet to commit.

Depends on Dxxxx.

Part of Txxxx.

Test Plan: CTS, sqlite multithreaded-locking tests, and regular sqlite tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx