verneuil/tracker: compare our snapshot with reality in test mode
Summary:
When built as a test VFS, compute a blake2b digest of the actual
db file, and compare that with a digest of the snapshotted data.

We use blake2b because it's well-known and obviously independent
of any potential issue in umash.

If that ever fails, panic.

Depends on Dxxxx.

Improves test coverage for Txxxx.

Test Plan: Sqlite tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx