verneuil/process_id: construct unique-ish process identifiers
Summary:
The problem with pids is that they are eventually recycled, so it can
be ambiguous whether an identifier refers to current process at that
pid, or an earlier one. Couple the pid with the birth tick for the
process, which should make collisions extremely unlikely (we'd have
to recycle the same pid within a single tick).

Will be used in temporary file names for snapshotting (Txxxx).

Test Plan: New unit tests.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx