verneuil/tracker: skip useless replication work more finely
Summary:
Now that we also publish to the "ready" buffer directory, we're only
done when that directory is up-to-date.  However, we can still skip
some of the snapshotting work when "staging" is already up-to-date as
well.

Depends on Dxxxx.

Lessens wasted work in concurrent snapshot publication for Txxxx.

Test Plan: Sqlite tests?  This is mostly about reading wait-free logic.

Reviewers: ahicks, sbahra, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx