verneuil/tracker: clean up scratch directory after publication
Summary:
Once we've updated the "ready" directory, there is no work left to do
until we release the sqlite read lock.  That means we can
unconditionally delete all temporary files and directory in the
"staging/scratch" subdirectory... as long as we make directories
disappear atomically: we don't want a rename(2) to spuriously succeed
in publishing a half-deleted subdirectory.

Improves crash cleanliness of the snapshotting logic in Txxxx.

Test Plan:
sqlite tests?  Mostly, this is about following the logic to confirm
that it's race-free; testing for POSIX race is hard.

Reviewers: wandrews, sbahra, ahicks

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx