verneuil/directory_schema: acquire fingerprint functions from tracker.rs
Summary:
The fingerprint logic is part of the directory format, so it makes
more sense in the schema file than in a module that happens to create
directories.

In preparation for checksumming on reads (Txxxx).

Test Plan: Just moving code around.  There is a new smoke test for the chunk list fingerprint.

Reviewers: wandrews, ahicks, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx