verneuil/copier: new component to asynchronously send data to blob stores
Summary:
The tracker / replication buffer combo is responsible for constructing
and atomically updating snapshots of consistent replication
data (chunks and directory) for a database file.

We just have to send that data somewhere, now that the tracker can
signal when it pushes newly consistent data.

Part of Txxxx.

Test Plan: Follow-up commits.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx