verneuil/copier: acquire and consume ready buffer directories
Summary:
This diff implements all the local filesystem manipulation to
atomically acquire and dump a directory to a remote blob store.
However, the actual copying logic is still stubbed out.  That's
the next step.

For the initial rollout, we prefer consistency over liveness or
availability, *for the replicated data*. The plain sqlite db file
always makes progress as usual (CAP is irrelevant for a single-node
system), but will run ahead of replication when replication fails.
That's safe because the replication protocol bounds the amount of
state it keeps around to a small multiple of the base DB file.

It's also correct because the replication subsystem does not provide
any liveness guarantee.  In fact, when something goes wrong with a
replication buffer's state, we expect it to remain stuck until the
directory is cleared. We'll want to improve visibility around that
failure mode (Txxxx).

Depends on Dxxxx, Dxxxx, Dxxxx.

Part of Txxxx.

Test Plan:
Existing tests.  We'll have more useful integration coverage
once the blob copying logic is in.

Reviewers: ahicks, wandrews

Reviewed By: ahicks, wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx