verneuil/vfs_ops: force a snapshot after every write transaction
Summary:
Sqlite will sometimes write to a db file with the lock held, but
without making any semantic change to the database's contents; for
example, that happens when a write transaction explicitly flushes its
(temporary?) dirty pages to disk.

We aim for bitwise identical replication, so catch those cases by
tracking write operations, and forcing a snapshot whenever sqlite
releases a write lock after performing writes, even if it claims the
transaction was a no-op.

This (rare) possibility also means that readers may find a replicated
change that has the same header as the current file, but different
contents.  That's why we only check that snapshots and reality match
after we have definitely written a fresh snapshot.  Our *test* invariant
is now that the replication data always matches the real file when
it's fresh... and the replication data is always fresh immediately
after a write transaction.

Part of Txxxx.

Test Plan: sqlite tests workingness is restored.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx