verneuil/replication_target: add optional endpoint field
Summary:
When we use a test region hosted on an ad hoc server (e.g., minio on
localhost), we want to specify a region that looks like a region
name (e.g., "minio-test"), and override that with an explicit
endpoint.

Add that field to the config JSON.  Next step: use it in `copier.rs`.

Improves testability of Txxxx.

Test Plan: We introduced this field for testing.  Follow-up commits.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx