verneuil/replication_buffer: separate fingerprint hashes with %2F
Documentation on sharding keys in S3 is a bit ambiguous, and some
documents imply that the sharding key only considers the "directory"
prefix in a blob name.  Our chunks' names are high-entropy hashes, so
that's a bit unfortunate.  Defensively work around that logic by
separating the primary and secondary hashes with a url-encoded forward
slash.

Avoids potential scaling issues in Txxxx.