verneuil/tracker: validate snapshot directory files
Summary:
We can't always assume that the directory files are up to date.
However, we *can* make sure that they make sense.  In particular,
we should be able to fetch all the chunks they refer to.

Depends on Dxxxx.

More test assertions for Txxxx.

Test Plan: sqlite tests still pass.

Reviewers: ahicks, sbahra, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx