verneuil/tracker: remove all staged chunks once published to "ready"
Summary:
If we successfully published a new "ready" directory, all chunks we
currently care about are in that directory; we can remove everything
from the staging chunks directory.

If we fail to publish a new directory, we still want to GC: we don't
want to let the staging directory grow an unbounded number of chunks
when replication is stuck.

Depends on Dxxxx, Dxxxx.

Avoid redundant storage and copies in Txxxx.

Test Plan: sqlite tests.

Reviewers: ahicks, sbahra, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx