verneuil/tracker: amortise chunk GC with randomised counting
Summary:
When replication falls behind, we want to remove obsolete chunk files,
chunks to which the directory file does not refer.  However, we do
want to keep useful chunks around.  We still have to scan those chunks
to realise that we don't want to delete them, so we end up wasting
time proportional to the number of referenced chunks, in the worst
case (the time we spend on chunk files we do delete can be attributed
as part of the time it took to create them).

Make sure to only trigger a GC from time to time, when the number
of files we have copied since the last GC is roughly equal to the
number of chunks we want to keep.

Fixes asymptotic complexity of Txxxx when replication gets stuck.

Test Plan:
Look at the code, glance at `staging` directories on disk: leaving
chunk files behind should be a quality issue, not correctness.

Reviewers: ahicks, sbahra, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx