verneuil/replication_buffer: populate .metadata in staging directories
Summary:
Copiers will want to opportunistically copy data from staging
directories to blob stores, when the DB is quiescent.  In order
to do that, we want need `.metadata` in there.

That file might be stale, so we update it unconditionally when we
forced an update to the replicated data (i.e., after a sqlite write
transaction).

Scaffolding for livelier replication in Txxxx.

Test Plan: sqlite tests don't crash; follow-up commit.

Reviewers: ahicks, sbahra, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx