verneuil/copier: improve liveness when the staging buffer is quiescent
Summary:
The change `Tracker` only updates the "ready" buffer when it has been
consumed by a copier.  That's great for correctness, and helps limit
the amount of work a `Tracker` must perform (we have a guarantee that
"ready" data eventually makes it to remote storage).

However, that's not ideal for liveness, since a copier that falls
behind can find a very stale "ready" buffer.

That's OK if writes are still happening frequently: the copier will
consume the stale ready buffer, the tracker will close another
transaction, and publish a new ready buffer.

That's not OK when writes don't happen anymore.  That's why the copier
now attempts to publish data from the staging directory, but only when
it's obviously quiescent.

The copier uniquely identifies all staged metadata files, consumes all
staged chunks, and then checks for any new chunk, in the staging or
ready buffers (directories).  If it doesn't find any, all metadata
files that haven't changed are safe to upload.

Otherwise, the staging buffer isn't quiescent, so we might have missed
some chunks (the tracker deletes staged chunks when it publishes a new
ready buffer)... but that's OK, because this means that a tracker has
or will soon publish an updated ready directory.

This synchronisation scheme can be seen as a hybrid of RCU and
sequence locking: the RCUed ready buffer guarantees copiers don't
starve, while the sequence locked staging buffer guarantees liveness
(freshness) when writers (`Tracker`s) are idle.

Depends on Dxxxx, Dxxxx, Dxxxx.

Improves the liveness guarantees of the replication logic in Txxxx.

Test Plan: sqlite tests, but mostly looking at the code.

Reviewers: sbahra, ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx