verneuil/tracker: always assert that the contents' length matches
Summary:
We can't always compare directories with the current state of the db,
but we *can* always make sure that they make sense.  We currently
check that we can find every chunk the directory proto refers to.
Let's also make sure that the sum of the chunks' lengths matches
the length in the proto.

Depends on Dxxxx.

Improves coverage for Txxxx.

Test Plan: Sqlite test pass with new assertions.

Reviewers: ahicks, wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx