verneuil/tracker: fast-path zero-filled chunks in internal tests
Summary:
Some sqlite tests generate very large sparse files.  The replication
logic handles them fine, especially with the deduplication that comes
naturally with content-addressed chunks.

The test-only checking logic, on the other hand, ends up fetching half
a million identical chunks from minio, and that takes a while.  We'll
instead special case the fingerprint for an all-zero chunk, and assume
we know its contents.

In theory, this does lose a bit of coverage, since we might fail to
upload zero chunks, but the replication logic doesn't treat the zero
chunk specially, so that seems unlikely to be a problem.  Moreover,
the read-side might also want to special-case the zero chunk and not
fetch it from remote storage.

Depends on Dxxxx.

Shortens test times for Txxxx.

Test Plan: Tests pass?

Reviewers: ahicks, wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx