verneuil/copier: no-op when there is no replication target
When the replication configuration doesn't yield any target, we don't
want to claim that we successfully copied everything to all
targets (i.e., to /dev/null).  Let's just abort out of the copy
functions.

This is safe, and the update protocol guarantees that the space
usage is still bounded.

Fixes a correctness hole before testing a new setup for Txxxx.