verneuil/tracker: simplify the noop detection logic when snapshotting
Summary:
The only fast path we can safely implement is detecting when the
current staged snapshot seems to match the current db file, and only
when we don't know for sure that we have just mutated the file.

Remove everything else, it's not useful.

Before adding more snapshotting complexity for Txxxx.

Test Plan: sqlite tests?

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx