verneuil/directory_schema: also track an external version (uu)id
Summary:
In some cases, sqlite will update a file's contents without
incrementing the version counter in the header; this can happen when
flushing the page cache, for example.  We always force a snapshot
after writes, so this can only be an issue when the writer crashes
between the sqlite commit and updating the snapshot in "staging."

That's a pretty small window, but let's try to avoid that by
storing a uuid in an xattr, when possible (xattrs are pretty old,
and are enabled by default in ext4 and xfs).

Also renumber the fields: no one is really persisting directory
protos yet.

Lowers the probability of spurious no-ops before finer change tracking (Txxxx).

Test Plan: sqlite tests. mptest in particular, since it's timing (performance) sensitive.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx