verneuil/tracker: stash the file's version id before we write to it
Summary:
We want to add incremental diffing, so we'll have to determine when
that optimisation isn't applicable, when the file and the current
snapshot have diverged.

This commit stashes the version id immediately after acquiring the
write lock, but doesn't use it for anything: merely recording the
version_id xattr can make sqlite's `mptest` flaky, if we do it too
inefficiently (e.g., with the xattr crate).

Depends on Dxxxx.

In preparation for incremental change tracking (Txxxx).

Test Plan: sqlite tests. mptest in particular, since it's timing (performance) sensitive.

Reviewers: ahicks, wandrews, sbahra

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx