verneuil/tracker: track changes incrementally
Summary:
Now that we can reliably tell when a snapshot is desynchronised
from the state of the db file when we entered the write transaction,
we can safely use a list of dirty chunks to drive replication.

After a write transaction, we check if the current staged snapshot
matches either the previous or the current db file state (if it
matches the current state, we still want to force a snapshot, just in
case the match is spurious).  If so, we can work off of that.

When in doubt, revert to the older logic and snapshot everything from
scratch.

Depends on Dxxxx.

Closes Txxxx: we now have incremental change tracking.

Test Plan: sqlite tests, with xattr disabled, and with change detection exclusively with xattrs.

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx