verneuil/replication_buffer: stricter permissions on replication dirs
We default to 0700, to avoid leaking information by default.  However,
we only apply this permission mask when we create the base replication
directory, so operators can easily widen the scope after the fact.

This default can also be more persistently loosened in the config JSON.

Marginally improves safety of Txxxx.