verneuil/replication_buffer: fast path redundant chunk writes
When the file for a given chunk already exists, we don't have
to try to populate it: the call will fail with EEXIST.

However, that's also not an error: all we want is for the chunk to
exist, it doesn't matter if we're the one who actually created it.

Improves performance of Txxxx.