verneuil/copier: pace calls to the remote store (e.g., S3)
Summary:
When we modeled replication costs, we assumed a bound on the rate of
PUT calls per second.  Enforce such a limit with a GCRA (leaky bucket)
limiter: we want to make sure we average out to not too many calls per
second (currently set to 2/second).  However, we also want to allow
bursts when there is work to do: we expect our write load to be
sparse, so let's make sure the data is copied ASAP when we do find
something to upload.

Closes Txxxx.

Test Plan: sqlite tests?

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx