verneuil/tracker: stage chunks incrementally as we write them out
Summary:
When sqlite is configured with 64 KB pages, we expect most writes
to be aligned to 64 KB and to span exactly 64 KB of data.  Detect
these easy cases, and immediately stage the chunks for replication.

Depends on Dxxxx.

Closes Txxxx.

Test Plan:
sqlite tests, with the page size overridden to 64KB (instead of 1KB).
A lot of tests failed, but we didn't trigger our internal consistency
checks.

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx