verneuil/copier: maintain set of active spool directories
Summary:
We use a bounded channel to send join/leave (add/remove directory)
messages to the worker.  The senders always block, so we know the
messages are delivered reliably and in order.

Now that we have this set around, we can run a background scan to
find any spool directory for which the `signal_ready_buffer` might
have failed (because the channel was full).

Depends on Dxxxx, Dxxxx.

In preparation for Txxxx.

Test Plan: sqlite tests.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx