verneuil/copier: improve fairness of background scans
Summary:
Try to make sure that `handle_spooling_directory` always does some
work: when we find an empty "ready" directory, always try to handle
the "staging" directory, without waiting for the governor's OK.

Internally, `handle_staging_directory` will still wait for the
governor, so rate limiting is still in effect.  The difference is
that we will actually sleep instead of bailing out.

Improves fairness of Txxxx.

Test Plan: Hard to test for fairness. Read the code?

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx