verneuil/copier: implement exponential backoff
Summary:
When we fail to connect to the blob store, or it responds with a 5xx,
we want to quickly try again.  Do that with a bounded exponential
backoff.

After 2 retries, we give up: the failure will be bubbled up to the
caller, which will leave the work behind for us to pick up at the
next background scan.

Depends on Dxxxx: it's easier to sleep in dedicated upload workers.

Improves the resilience of Txxxx to transient failures.

Test Plan: This logic isn't triggered in tests.  Examine the code :\

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx