verneuil/replication_buffer: allow TTL on scratch *files*
Summary:
We're about to add logic where *`Copier`s* create scratch files.  This
means a `Tracker` won't be able to assume that all scratch data is
useless after a successful snapshot.

We will still make sure that code cleans up after itself and deals
gracefully with scratch files that have disappeared, but not deleting
files that are most probably still in use is good for progress when
everything is in a good state.

In preparation for stashing copied meta directory files (Txxxx).

Test Plan:
Existing tests, but we don't exercise that logic in tests (we
always clean up gracefully).

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx