verneuil/tracker: store ctime in the directory proto
Summary:
A file's ctime can never be used for correctness, but it is useful for
reporting, e.g., on the replication lag.

Also renumber the fields one last (hopefully) time, while this is
still all preprod.  It's easier to hackily skip over the potentially
large `chunks` array if that field has a higher tag value than
everything we want to parse; 15 is the highest protobuf tag that can
be encoded in one byte, so use that.

In preparation for replication lag stats (Txxxx).

Test Plan: sqlite tests.

Reviewers: ahicks, wandrews

Reviewed By: ahicks, wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx