NFC verneuil/copier: simplify `with_spool_path`
We get a fresh copier via `get_global_copier` or `clone`, and only use
`with_spool_path` to update a `Copier` with a new `spool_path`.

Also, when we make a copier with a new spool path, and that path
happens to be the same as the new one, we don't have to twiddle
refcounts.

Noticed while working on replication lag reporting (Txxxx).