verneuil/tracker: forcibly set a version xattr after an initial read txn
When we open a new db file that does not have a version xattr, but
never write to that file, we will keep observing an empty xattr, and
enter the slow replication path.

Try to avoid that by replacing the empty xattr with a fresh version
uuid.

Noticed while working on Txxxx.