verneuil/copier: acquire local file lock around directory blob uploads
If we let multiple copiers concurrently upload directory blobs for the
same spooling directory, a slow upload could end up overwriting a more
recent directory... which would make the snapshot go back in time (bad).

We already enforce mutual exclusion internally, with `Mutex`es, so the
only way for the lock acquisition to fail is if there are multiple
processes copying from the same directory.  That's supported, but not
currently exercised in production.

Improves the consistency guarantees of the replication code (Txxxx).