verneuil/tracker: increase probability of a chunk GC
We now expect to trigger a GC whenever we have useless chunks on the
order of 25% of useful ones.  This makes it likely (> 98%) that we
will trigger a GC before useless chunks outnumber useful ones, and
highly unlikely (< 2e-7) that we will have 3 times as many useless as
useful chunks.

Ref Txxxx.