verneuil/copier: regularly "touch" useful chunks
Summary:
We can touch a blob in S3 by copying it over itself (as long as we
also set the request to replace the metadata).  Do that for a small
random fraction of the chunk blobs referenced in the most recently
uploaded directory for each replicated db file, whenever we look at
that file's spooling directory.

When we observe a definite failure (e.g., 404), we must assume
something's wrong with the replication data we uploaded to the remote
blob store.  That's why we eat the error, but make sure the next
update to the spooling directory starts from scratch, instead of
being computed incrementally: we change the db file's version tag,
and erase the staged directory proto file.

Closes Txxxx; it's now possible to use Verneuil with retention rules.

Test Plan:
sqlite tests, with the touch coverage fraction hardcoded to 100%.
Also injected failures, and confirmed that we do clear the pending state.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx