verneuil/vfs: sync with fsync instead of fdatasync
We update xattrs on sqlite db files in the middle of write
transactions, and it's not obvious that they are covered by fdatasync.

We don't *currently* drag state across reboots, but that could change,
e.g., if we restore the initial replication state from remote storage.
Let's just use fsync.

Ref Txxxx.