verneuil/copier: improve staged directory handle on FSes without btime
We don't expect to run on filesystems without btime, but if that's
somehow not available, we should still try to behave nicely.

We will for uploads regardless of what the cache says:

1. with a small random probability
2. whenever we find chunks to upload

This guarantees that we will eventually upload a staged directory
file, even if we think we have already seen it.  That in turn lets
us simplify `FileIdentifier::equal_except_ctime` by allowing more
false positives.

Ref Txxxx.