verneuil/copier: consider sqlite header fprint in staleness reports
When the source sqlite db file's ctime differs from the replicated
data's, but the sqlite header fingerprints match, we're almost surely
looking at a false positive: something changed the source file's ctime
(e.g., chmod), but the contents are semantically the same.  In fact,
in most cases the contents are actually unchanged.

Either way, we don't really want to consider a replica stale when the
header fingerprints match.  We still want to log about such false
positives, just in case, but we will now do so at the info level.

Closes Txxxx, now that the staleness report is less noisy.