verneuil/tracker: stage no-op chunks for upload
We randomly spot check a fraction of clean chunks to make sure they
actually match our current snapshot.  When we do that, stage them
for upload to remote storage: we might as well use the opportunity
to repair anything that might be off with the remote data.

Improves the resilience of Txxxx.