verneuil/copier: make progress even when the copy lock is taken
The meta copy lock only exists to prevent races where a slow upload of
a stale directory proto blob overwrites a quicker upload of a more
recent proto blob.  It's not necessary for correctness, and, even if
such a race happens, we will eventually recover as long as copiers can
make progress: copiers always (eventually) try to upload the directory
blob in staging, if there is one.

What's important is thus to make the race unlikely, without ever
letting a lock owner block progress.  We do that by sleeping for a bit
on lock acquisition failure, but always letting the copier upload
directory blobs, even if it could not acquire the copy lock.

Ref Txxxx.