verneuil/copier: try to queue stale spooling directories for copy
Summary:
We sometimes detect that replicated data is far behind the source
file.  This seems to mostly happen after a surge of database updates
temporarily overflows the change-triggered `ready_buffers` queue.
Eventually, a background scan will detect the spooled data and upload
it, but, when a process is replicating a couple hundred databases,
this could take a while.

When the copier backend thread detects such a replication lag, attempt
to fast-track the spooling directory for uploads.  We don't want to
block regular change-triggered replication with these fast-tracking
requests, so use a dedicated channel.

Accelerates recovery from queue overflow in Verneuil (Txxxx).

Test Plan: Deploy to yolo, where we observe such spikes every day or two.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx