verneuil/lib: let users set the default config in an environment variable
Summary:
When we don't have a config pointer, the C code will read
`VERNEUIL_CONFIG`.  If that environment variable has a non-empty
value, we will attempt to parse it as a JSON config, or as a path to a
JSON config file if the string is of the form "@/path/to/file.json"

Combined with the new logic to parse a configuration string, this
change makes it easier for tooling to let users configure Verneuil at
runtime.

In preparation for read replicas (Txxxx).

Test Plan:
Updated sqlite tests, with minio enabled (which uses the "@path" logic),
and disabled (which uses an inline JSON string).

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx