verneuilctl-restore: also accept a hostname / path pair
Summary:
The command line now accepts either the path to a manifest file, or
the path of the replicated db and the hostname.  If the latter are
provided, verneuilctl will attempt to download the corresponding
blob from remote storage before restoring the contents.

```
$ ./verneuilctl -c @./verneuil.json restore -s /etc/coronerd/coronerd.db -o coronerd.db
$ sudo md5sum /etc/coronerd/coronerd.db
89874f7f98e0794d419309d795b17b84  /etc/coronerd/coronerd.db
$ md5sum coronerd.db
89874f7f98e0794d419309d795b17b84  coronerd.db
```

Depends on Dxxxx.

Improved tooling for Verneuil (Txxxx).

Test Plan: Manual.

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Differential Revision: https://foo.backtrace.io/Dxxxx