verneuil: let C callers fetch manifests with `verneuil_replication_info_for_manifest`
Summary:
We have a function that attempts to fetch the local manifest for a
given db path.  We can populate the same struct, but using remote
information this time.

Ref Txxxx.

Test Plan:
The rust logic has already been tested. This is all type munging.  We
don't have a deep C integration, it just makes sense to have this
function.

Reviewers: ahicks, wandrews

Reviewed By: wandrews

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx