verneuil/replication_buffer: make slash replacement invertible
Percent-encode control characters, `%`, and `#`, then replace `/` with
`#`.  That's invertible, and equivalent to the old logic where we only
replaced `/` with `#` for all paths we care about.  We don't currently
rely on invertibility, but it doesn't hurt to guarantee the property.

We don't want to merely percent-encode slashes away because the result
is harder to read when listing directories, for local administration.

Ref Txxxx.