verneuil/replication_buffer: always clean up stale replication directories
The encoding is now unambiguous, so we know that any replication
directory for a different inode than the one we just had is wrong;
we can delete those.

It's also possible that the inode we opened has since been unlinked,
but you're not supposed to do that while sqlite is using your files.

Ref Txxxx.