verneuil/copier: prioritise edge / lag triggered copies
Only make half (rounded up) the copy workers look for maintenance
work.  The fraction is arbitrary, but the idea is that we want to
guarantee some workers look for background maintenance work, but also
reserve some workers to only pull edge- and lag- triggered work: we
know those are important.

Ref Txxxx.