verneuil/copier: issue a GET before every PUT
Summary:
Rumour has it S3 likes to buffer PUTs for a long time, and that can
result in bad performance when overwriting a PUT that is still
queued up.  Apparently, we can force things out of that buffer by
first issuing a GET for the blob we're about to (over)write.

Depends on Dxxxx.

May help reduce the occurrence of slow requests noted in Txxxx.

Test Plan: Sqlite tests.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Differential Revision: https://foo.backtrace.io/Dxxxx