verneuil/copier: rename "ready" directory before consuming it
Summary:
When snapshotting a new "ready" directory, we will first try to rename
it to "consuming": this frees up the directory for `Tracker`s to
publish a fresher snapshot while the copier is working.

If the rename succeeds, fails because "ready" does not exist anymore,
or fails because "consuming" is non-empty, then try to snapshot
"consuming" and copy / delete files from there.  Finally, once we're
done with the "consuming" directory, delete it; there's no need to
clean up the "ready" directory because the rename does that already.

Of course, this now means that we must check for the existence of a
non-empty "ready" *or* "consuming" directory, when we used to only
check for a non-empty "ready" directory.

Depends on Dxxxx, Dxxxx.

Closes Txxxx.

Test Plan: sqlite tests.

Reviewers: ahicks, wandrews

Reviewed By: ahicks

Maniphest Tasks: Txxxx

Differential Revision: https://foo.backtrace.io/Dxxxx