verneuil/tracker: stage less useless work
and consistently look for unexpectedly dirty chunks.

1. When a chunk matches what we expect (potentially after applying
   a write transaction), we don't want to stage it for uploads: either
   the chunk is clean and doesn't need copying, or it was already
   staged during the write.  Either way, avoid redundant uploads.

2. When a chunk doesn't have the value we expect after incremental
   staging, trigger a full re-scan: if something's off with our
   incremental staging, the whole thing is probably wrong.

We have yet to trigger a full rescan, so change #2 is probably a no-op
in practice (but does help resilience).

Change #1, on the other hand, is important for replication lag: when
we observe a lot of back-to-back write transactions that hit the same
chunks, we can end up staging a lot of useless chunks, and that leads
to a replication hiccup when a copy worker gets to the spool directory.

Noticed while investigating Txxxx on > 500 MB sqlite dbs.

TESTED=sqlite tests, with SNAPSHOT_GRANULARITY = 2^{9..13, 16}.