verneuil/tracker: avoid `/proc/self/fd/` in consistency check
We want to compare a tracker's snapshot with the contents of the
source replicated file.  The current code re-opens that file via
`/proc`.  That's ideal for isolation, but inspecting the VFS code
shows that it deals fine with dup/close (important with fork), and
that it never uses the file's current offset.

This makes it safe to directly reuse the tracker's file object, and
that's preferable to linux-specific (test-only) correctness checks.

Part of Txxxx.