verneuil/replication_buffer: version spooling directory names
If we switch to incompatible structures or contents for spooled
replication format, we'll want to denote that by changing the version
tag: in the common case, only one version of Verneuil is running at a
time, and this will make version changes appear as stale or a lack of
replication data.  We already have to deal with that.

In preparation for Txxxx.