verneuil/tracker: determinicise chunk GC in the worst case
We want to rely on memory-less randomised counting to trigger
amortised GC of useless chunks: that's safer with multiple processes,
or with short-lived processes.  However, we can still clamp the time
between GCs in long-lived processes without hurting our
non-synchronising properties too badly: we just have to make sure
the deterministic trigger is usually preempted by a random trigger.

That's what we do here, by counting the number of chunks published
since the last chunk GC, but only forcing a GC when that count
exceeds *twice* the number of chunks in the source db.  The random
trigger aims to trigger every 1/4th of the source db's chunk count,
so we should only hit the deterministic condition in a window of
2x chunk count publications less than ~ exp(-8) < 0.05% of the time.

Ref Txxxx.